♻️ refactor: activity 순수 함수, pipeline, 컴포넌트 분리
