+++ client/src/components/profile/header/ui/ActivityGraph/DayLabels.tsx
@@ -1,0 +1,7 @@
+export const DayLabels = () => (
+  <div className="flex flex-col justify-between text-xs text-gray-400 pr-2">
+    <span className="translate-y-2">Mon</span>
+    <span>Wed</span>
+    <span className="-translate-y-2">Fri</span>
+  </div>
+);
